+++ src/main/java/com/daruda/darudaserver/domain/community/dto/res/BoardRes.java
@@ -41,4 +41,5 @@ public class BoardRes {
                 .commentCount(commentCount)
                 .build();
     }
+
 }
